# openapi/nexusleo.openapi.json
{
  "components": {
    "schemas": {
      "ActivityLogCreate": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "entry_type": {
            "title": "Entry Type",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "entry_type",
          "text"
        ],
        "title": "ActivityLogCreate",
        "type": "object"
      },
      "ActivityLogOut": {
        "properties": {
          "case_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_type": {
            "title": "Entry Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata_json": {
            "additionalProperties": true,
            "title": "Metadata Json",
            "type": "object"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "shift_id": {
            "format": "uuid",
            "title": "Shift Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "shift_id",
          "occurred_at",
          "entry_type",
          "text",
          "metadata_json",
          "created_at"
        ],
        "title": "ActivityLogOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestRequest": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "IngestResult": {
        "properties": {
          "audit_events_created": {
            "title": "Audit Events Created",
            "type": "integer"
          },
          "claim_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Claim Ids",
            "type": "array"
          },
          "claims_created": {
            "title": "Claims Created",
            "type": "integer"
          },
          "confidence_created": {
            "title": "Confidence Created",
            "type": "integer"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "entities_created": {
            "title": "Entities Created",
            "type": "integer"
          },
          "evidence_links_created": {
            "title": "Evidence Links Created",
            "type": "integer"
          },
          "mentions_created": {
            "title": "Mentions Created",
            "type": "integer"
          }
        },
        "required": [
          "document_id",
          "mentions_created",
          "entities_created",
          "claims_created",
          "evidence_links_created",
          "confidence_created",
          "audit_events_created",
          "claim_ids"
        ],
        "title": "IngestResult",
        "type": "object"
      },
...
      "ShiftEndResponse": {
        "properties": {
          "ended_at": {
            "format": "date-time",
            "title": "Ended At",
            "type": "string"
          },
          "shift_id": {
            "format": "uuid",
            "title": "Shift Id",
            "type": "string"
          }
        },
        "required": [
          "shift_id",
          "ended_at"
        ],
        "title": "ShiftEndResponse",
        "type": "object"
      },
      "ShiftStartRequest": {
        "properties": {
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "ShiftStartRequest",
        "type": "object"
      },
      "ShiftStartResponse": {
        "properties": {
          "shift_id": {
            "format": "uuid",
            "title": "Shift Id",
            "type": "string"
          }
        },
        "required": [
          "shift_id"
        ],
        "title": "ShiftStartResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NexusLEO API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cases/{case_id}/ingest": {
      "post": {
        "operationId": "ingest_cases__case_id__ingest_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/shifts/start": {
      "post": {
        "operationId": "start_shift_shifts_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Shift",
        "tags": [
          "shifts"
        ]
      }
    },
    "/shifts/{shift_id}/end": {
      "post": {
        "operationId": "end_shift_shifts__shift_id__end_post",
        "parameters": [
          {
            "in": "path",
            "name": "shift_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shift Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftEndResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Shift",
        "tags": [
          "shifts"
        ]
      }
    },
    "/shifts/{shift_id}/log": {
      "post": {
        "operationId": "log_activity_shifts__shift_id__log_post",
        "parameters": [
          {
            "in": "path",
            "name": "shift_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shift Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityLogCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityLogOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Activity",
        "tags": [
          "shifts"
        ]
      }
    },
    "/shifts/{shift_id}/timeline": {
      "get": {
        "operationId": "timeline_shifts__shift_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "shift_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shift Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityLogOut"
                  },
                  "title": "Response Timeline Shifts  Shift Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timeline",
        "tags": [
          "shifts"
        ]
      }
    },
...
  }
}